requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]>=4.3
certifi
scikit-learn